7.2.0

+++ src/Umbraco.Core/Persistence/Factories/ServerRegistrationFactory.cs
@@ -28,7 +28,7 @@ namespace Umbraco.Core.Persistence.Factories
                     ComputerName = entity.ComputerName
                 };
             if (entity.HasIdentity)
-                dto.Id = short.Parse(entity.Id.ToString(CultureInfo.InvariantCulture));
+                dto.Id = int.Parse(entity.Id.ToString(CultureInfo.InvariantCulture));
 
             return dto;
         }
